requirements.txt
python-dotenv~=0.19.2
transformers~=4.12.5